Updated copyright notice to 2016; fixed formatting

+++ kse/src/net/sf/keystore_explorer/utilities/io/UnsignedUtil.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2004 - 2013 Wayne Grant
- *           2013 - 2015 Kai Kramer
+ *           2013 - 2016 Kai Kramer
  *
  * This file is part of KeyStore Explorer.
  *
@@ -26,7 +26,7 @@ import java.nio.ByteBuffer;
  * object. As Java has no concept of unsigned values the next biggest type is
  * used when getting or putting values to the ByteBuffer, e.g. when putting a
  * byte we supply a short, when we get an integer we actually get a long.
- * 
+ *
  */
 public class UnsignedUtil {
 	private UnsignedUtil() {
@@ -34,7 +34,7 @@ public class UnsignedUtil {
 
 	/**
 	 * Get an unsigned byte from the byte buffer.
-	 * 
+	 *
 	 * @param bb
 	 *            Byte buffer
 	 * @return Unsigned byte stored in a short to retain sign
@@ -45,7 +45,7 @@ public class UnsignedUtil {
 
 	/**
 	 * Put a byte into the byte buffer unsigned.
-	 * 
+	 *
 	 * @param bb
 	 *            Byte buffer
 	 * @param value
@@ -57,7 +57,7 @@ public class UnsignedUtil {
 
 	/**
 	 * Get an unsigned short from the byte buffer.
-	 * 
+	 *
 	 * @param bb
 	 *            Byte buffer
 	 * @return Unsigned short stored in an int to retain sign
@@ -68,7 +68,7 @@ public class UnsignedUtil {
 
 	/**
 	 * Put a short into the byte buffer unsigned.
-	 * 
+	 *
 	 * @param bb
 	 *            Byte buffer
 	 * @param value
@@ -80,18 +80,18 @@ public class UnsignedUtil {
 
 	/**
 	 * Get an unsigned int from the byte buffer.
-	 * 
+	 *
 	 * @param bb
 	 *            Byte buffer
 	 * @return Unsigned int stored in a long to retain sign
 	 */
 	public static long getInt(ByteBuffer bb) {
-		return ((long) bb.getInt() & 0xffffffffL);
+		return (bb.getInt() & 0xffffffffL);
 	}
 
 	/**
 	 * Put an int into the byte buffer unsigned.
-	 * 
+	 *
 	 * @param bb
 	 *            Byte buffer
 	 * @param value
